# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "To do API",
        "version": "1.0.0"
    },
    "paths": {
        "/api/v1/todo/lists": {
            "get": {
                "tags": [
                    "To do List"
                ],
                "summary": "Get all todo lists.",
                "description": "Search for cars for hire and for sale.",
                "operationId": "v1TodoLists",
                "responses": {
                    "200": {
                        "description": "Lists of tasks",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "boolean",
                                            "example": "false"
                                        },
                                        "status_code": {
                                            "type": "integer",
                                            "example": "200"
                                        },
                                        "status": {
                                            "type": "string",
                                            "example": "OK"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Lists of tasks"
                                        },
                                        "data": {
                                            "properties": {
                                                "total": {
                                                    "type": "integer",
                                                    "example": "2"
                                                },
                                                "query": {
                                                    "properties": {
                                                        "name": {
                                                            "type": "string",
                                                            "example": "Task name"
                                                        },
                                                        "activity": {
                                                            "type": "string",
                                                            "example": "Task activity"
                                                        },
                                                        "data": {
                                                            "type": "string",
                                                            "example": "2022/03/05"
                                                        },
                                                        "status": {
                                                            "type": "string",
                                                            "example": "failed"
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/todo/create": {
            "post": {
                "tags": [
                    "To do List"
                ],
                "summary": "Create a new Task",
                "description": "Create a new Task",
                "operationId": "v1TodoCreate",
                "requestBody": {
                    "description": "Task informations",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "activity",
                                    "date"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "task name"
                                    },
                                    "activity": {
                                        "type": "string",
                                        "example": "Description"
                                    },
                                    "date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "Activity date"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "422": {
                        "description": "An unknown error has occurred !!! Please try again later",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "boolean",
                                            "example": "true"
                                        },
                                        "status_code": {
                                            "type": "string",
                                            "example": "422"
                                        },
                                        "status": {
                                            "type": "string",
                                            "example": "Unprocessable Entity"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "An unknown error has occurred !!! Please try again later"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "Task added",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "boolean",
                                            "example": "false"
                                        },
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "status_code": {
                                            "type": "integer",
                                            "example": "200"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Task added."
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": "2"
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "Name task"
                                                },
                                                "activity": {
                                                    "type": "string",
                                                    "example": "Activity description"
                                                },
                                                "status": {
                                                    "type": "string",
                                                    "example": "completed"
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "example": "2022/02/12"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/todo/completed": {
            "post": {
                "tags": [
                    "To do List"
                ],
                "summary": "Completed Task",
                "description": "Completed Task",
                "operationId": "v1TodoCompleted",
                "requestBody": {
                    "description": "Task informations",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "id"
                                ],
                                "properties": {
                                    "id": {
                                        "type": "integer",
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "422": {
                        "description": "An unknown error has occurred !!! Please try again later",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "boolean",
                                            "example": "true"
                                        },
                                        "status_code": {
                                            "type": "string",
                                            "example": "422"
                                        },
                                        "status": {
                                            "type": "string",
                                            "example": "Unprocessable Entity"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "An unknown error has occurred !!! Please try again later"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "Completed task",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "boolean",
                                            "example": "false"
                                        },
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "status_code": {
                                            "type": "integer",
                                            "example": "200"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Task completed."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/todo/delete": {
            "post": {
                "tags": [
                    "To do List"
                ],
                "summary": "Completed Task",
                "description": "Delete Task",
                "operationId": "v1TodoDelete",
                "requestBody": {
                    "description": "Task informations",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "id"
                                ],
                                "properties": {
                                    "id": {
                                        "type": "integer",
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "422": {
                        "description": "An unknown error has occurred !!! Please try again later",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "boolean",
                                            "example": "true"
                                        },
                                        "status_code": {
                                            "type": "string",
                                            "example": "422"
                                        },
                                        "status": {
                                            "type": "string",
                                            "example": "Unprocessable Entity"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "An unknown error has occurred !!! Please try again later"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "Completed task",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "boolean",
                                            "example": "false"
                                        },
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "status_code": {
                                            "type": "integer",
                                            "example": "200"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Task delete."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "security": [
        []
    ]
}